+++ test/sentry_test.dart
@@ -11,9 +11,212 @@ import 'package:test/test.dart';
 import 'test_utils.dart';
 
 const String _testDsn = 'https://public:secret@sentry.example.com/1';
-
+const String _testDsnWithoutSecret = 'https://public@sentry.example.com/1';
+const String _testDsnWithPath =
+    'https://public:secret@sentry.example.com/path/1';
+const String _testDsnWithPort =
+    'https://public:secret@sentry.example.com:8888/1';
 void main() {
   group('$SentryClient', () {
+    test('can parse DSN', () async {
+      final SentryClient client = new SentryClient(dsn: _testDsn);
+      expect(client.dsnUri, Uri.parse(_testDsn));
+      expect(client.postUri, 'https://sentry.example.com/api/1/store/');
+      expect(client.publicKey, 'public');
+      expect(client.secretKey, 'secret');
+      expect(client.projectId, '1');
+      await client.close();
+    });
+
+    test('can parse DSN without secret', () async {
+      final SentryClient client = new SentryClient(dsn: _testDsnWithoutSecret);
+      expect(client.dsnUri, Uri.parse(_testDsnWithoutSecret));
+      expect(client.postUri, 'https://sentry.example.com/api/1/store/');
+      expect(client.publicKey, 'public');
+      expect(client.secretKey, null);
+      expect(client.projectId, '1');
+      await client.close();
+    });
+
+    test('can parse DSN with path', () async {
+      final SentryClient client = new SentryClient(dsn: _testDsnWithPath);
+      expect(client.dsnUri, Uri.parse(_testDsnWithPath));
+      expect(client.postUri, 'https://sentry.example.com/path/api/1/store/');
+      expect(client.publicKey, 'public');
+      expect(client.secretKey, 'secret');
+      expect(client.projectId, '1');
+      await client.close();
+    });
+    test('can parse DSN with port', () async {
+      final SentryClient client = new SentryClient(dsn: _testDsnWithPort);
+      expect(client.dsnUri, Uri.parse(_testDsnWithPort));
+      expect(client.postUri, 'https://sentry.example.com:8888/api/1/store/');
+      expect(client.publicKey, 'public');
+      expect(client.secretKey, 'secret');
+      expect(client.projectId, '1');
+      await client.close();
+    });
+    test('sends client auth header without secret', () async {
+      final MockClient httpMock = new MockClient();
+      final ClockProvider fakeClockProvider =
+          () => new DateTime.utc(2017, 1, 2);
+
+      Map<String, String> headers;
+
+      httpMock.answerWith((Invocation invocation) async {
+        if (invocation.memberName == #close) {
+          return null;
+        }
+        if (invocation.memberName == #post) {
+          headers = invocation.namedArguments[#headers];
+          return new Response('{"id": "test-event-id"}', 200);
+        }
+        fail('Unexpected invocation of ${invocation.memberName} in HttpMock');
+      });
+
+      final SentryClient client = new SentryClient(
+        dsn: _testDsnWithoutSecret,
+        httpClient: httpMock,
+        clock: fakeClockProvider,
+        compressPayload: false,
+        uuidGenerator: () => 'X' * 32,
+        environmentAttributes: const Event(
+          serverName: 'test.server.com',
+          release: '1.2.3',
+          environment: 'staging',
+        ),
+      );
+
+      try {
+        throw new ArgumentError('Test error');
+      } catch (error, stackTrace) {
+        final SentryResponse response = await client.captureException(
+            exception: error, stackTrace: stackTrace);
+        expect(response.isSuccessful, true);
+        expect(response.eventId, 'test-event-id');
+        expect(response.error, null);
+      }
+
+      final Map<String, String> expectedHeaders = <String, String>{
+        'User-Agent': '$sdkName/$sdkVersion',
+        'Content-Type': 'application/json',
+        'X-Sentry-Auth': 'Sentry sentry_version=6, '
+            'sentry_client=${SentryClient.sentryClient}, '
+            'sentry_timestamp=${fakeClockProvider().millisecondsSinceEpoch}, '
+            'sentry_key=public',
+      };
+
+      expect(headers, expectedHeaders);
+
+      await client.close();
+    });
+
+    testCaptureException(bool compressPayload) async {
+      final MockClient httpMock = new MockClient();
+      final ClockProvider fakeClockProvider =
+          () => new DateTime.utc(2017, 1, 2);
+
+      String postUri;
+      Map<String, String> headers;
+      List<int> body;
+      httpMock.answerWith((Invocation invocation) async {
+        if (invocation.memberName == #close) {
+          return null;
+        }
+        if (invocation.memberName == #post) {
+          postUri = invocation.positionalArguments.single;
+          headers = invocation.namedArguments[#headers];
+          body = invocation.namedArguments[#body];
+          return new Response('{"id": "test-event-id"}', 200);
+        }
+        fail('Unexpected invocation of ${invocation.memberName} in HttpMock');
+      });
+
+      final SentryClient client = new SentryClient(
+        dsn: _testDsn,
+        httpClient: httpMock,
+        clock: fakeClockProvider,
+        uuidGenerator: () => 'X' * 32,
+        compressPayload: compressPayload,
+        environmentAttributes: const Event(
+          serverName: 'test.server.com',
+          release: '1.2.3',
+          environment: 'staging',
+        ),
+      );
+
+      try {
+        throw new ArgumentError('Test error');
+      } catch (error, stackTrace) {
+        final SentryResponse response = await client.captureException(
+            exception: error, stackTrace: stackTrace);
+        expect(response.isSuccessful, true);
+        expect(response.eventId, 'test-event-id');
+        expect(response.error, null);
+      }
+
+      expect(postUri, client.postUri);
+
+      final Map<String, String> expectedHeaders = <String, String>{
+        'User-Agent': '$sdkName/$sdkVersion',
+        'Content-Type': 'application/json',
+        'X-Sentry-Auth': 'Sentry sentry_version=6, '
+            'sentry_client=${SentryClient.sentryClient}, '
+            'sentry_timestamp=${fakeClockProvider().millisecondsSinceEpoch}, '
+            'sentry_key=public, '
+            'sentry_secret=secret',
+      };
+
+      if (compressPayload) expectedHeaders['Content-Encoding'] = 'gzip';
+
+      expect(headers, expectedHeaders);
+
+      Map<String, dynamic> data;
+      if (compressPayload) {
+        data = json.decode(utf8.decode(gzip.decode(body)));
+      } else {
+        data = json.decode(utf8.decode(body));
+      }
+      final Map<String, dynamic> stacktrace = data.remove('stacktrace');
+      expect(stacktrace['frames'], const TypeMatcher<List>());
+      expect(stacktrace['frames'], isNotEmpty);
+
+      final Map<String, dynamic> topFrame =
+          (stacktrace['frames'] as Iterable<dynamic>).last;
+      expect(topFrame.keys,
+          <String>['abs_path', 'function', 'lineno', 'in_app', 'filename']);
+      expect(topFrame['abs_path'], 'sentry_test.dart');
+      expect(topFrame['function'], 'main.<fn>.testCaptureException');
+      expect(topFrame['lineno'], greaterThan(0));
+      expect(topFrame['in_app'], true);
+      expect(topFrame['filename'], 'sentry_test.dart');
+
+      expect(data, {
+        'project': '1',
+        'event_id': 'X' * 32,
+        'timestamp': '2017-01-02T00:00:00',
+        'platform': 'dart',
+        'exception': [
+          {'type': 'ArgumentError', 'value': 'Invalid argument(s): Test error'}
+        ],
+        'sdk': {'version': sdkVersion, 'name': 'dart'},
+        'logger': SentryClient.defaultLoggerName,
+        'server_name': 'test.server.com',
+        'release': '1.2.3',
+        'environment': 'staging',
+      });
+
+      await client.close();
+    }
+
+    test('sends an exception report (compressed)', () async {
+      await testCaptureException(true);
+    });
+
+    test('sends an exception report (uncompressed)', () async {
+      await testCaptureException(false);
+    });
+
     test('reads error message from the x-sentry-error header', () async {
       final ClockProvider fakeClockProvider =
           () => new DateTime.utc(2017, 1, 2);
@@ -117,6 +320,22 @@ void main() {
   });
 
   group('$Event', () {
+    test('$Breadcrumb serializes', () {
+      expect(
+        Breadcrumb(
+          "example log",
+          DateTime.utc(2019),
+          level: SeverityLevel.debug,
+          category: "test",
+        ).toJson(),
+        <String, dynamic>{
+          'timestamp': '2019-01-01T00:00:00',
+          'message': 'example log',
+          'category': 'test',
+          'level': 'debug',
+        },
+      );
+    });
     test('serializes to JSON', () {
       final user = new User(
           id: "user_id",
@@ -124,9 +343,16 @@ void main() {
           email: "email@email.com",
           ipAddress: "127.0.0.1",
           extras: {"foo": "bar"});
+
+      final breadcrumbs = [
+        Breadcrumb("test log", DateTime.utc(2019),
+            level: SeverityLevel.debug, category: "test"),
+      ];
+
       expect(
         new Event(
           message: 'test-message',
+          transaction: '/test/1',
           exception: new StateError('test-error'),
           level: SeverityLevel.debug,
           culprit: 'Professor Moriarty',
@@ -140,10 +366,12 @@ void main() {
           },
           fingerprint: <String>[Event.defaultFingerprint, 'foo'],
           userContext: user,
+          breadcrumbs: breadcrumbs,
         ).toJson(),
         <String, dynamic>{
           'sdk': {'version': sdkVersion, 'name': 'dart'},
           'message': 'test-message',
+          'transaction': '/test/1',
           'exception': [
             {'type': 'StateError', 'value': 'Bad state: test-error'}
           ],
@@ -158,6 +386,16 @@ void main() {
             'email': 'email@email.com',
             'ip_address': '127.0.0.1',
             'extras': {'foo': 'bar'}
+          },
+          'breadcrumbs': {
+            'values': [
+              {
+                'timestamp': '2019-01-01T00:00:00',
+                'message': 'test log',
+                'category': 'test',
+                'level': 'debug',
+              },
+            ]
           },
         },
       );