+++ tool/presubmit.sh
@@ -5,5 +5,6 @@ set -x
 
 pub get
 dartanalyzer --fatal-warnings ./
-pub run test --platform vm --platform chrome
+pub run test --platform vm
+./tool/dart2_test.sh
 dartfmt -n --set-exit-if-changed ./